+++ src/biokbase/narrative/__init__.py
@@ -1,7 +1,7 @@
 __all__ = ['magics', 'mongonbmanager', 'ws_util', 'common', 'kbasewsmanager', 'services']
 
 from semantic_version import Version
-__version__ = Version("2.0.4")
+__version__ = Version("2.0.7")
 version = lambda: __version__
 
 # if run directly: